增加赎回提货卡;bugfix

+++ equity/src/main/java/com/fengchao/equity/service/impl/CardInfoServiceImpl.java
@@ -81,7 +81,7 @@ public class CardInfoServiceImpl implements CardInfoService {
             throw new EquityException(MyErrorEnum.CARD_INFO_NOT_FOUND);
         }
 
-        Integer updateStatus = (int)bean.getStatus();
+        Short updateStatus = bean.getStatus();
         if (null != updateStatus){
             Integer status = (int)record.getStatus();
             if (CardInfoStatusEnum.INVALID.getCode().equals(status)) {
